+++ backend/cwa_lib/pages/su_manage_vdbs.py
@@ -7,7 +7,7 @@ from cwa_lib.pydantic_schemas.generic_table import (
     SelectOption, ColumnType, 
     TableOptions, TableQuery, TableCreateRowResult, TableUpdateRowResult, TableDeleteRowResult
 )
-from cwa_lib.pages import get_qc_to_with_group_id_name
+from cwa_lib.pages import get_qc_to_with_user_id_name_group_id_name
 from cwa_lib.pydantic_schemas.su_manage_vdbs import SuManageVDBsQueryResult, SuManageVDBsCreate, SuManageVDBsUpdate
 
 
@@ -49,8 +49,9 @@ class SuManageVDBsTable:
             deleted: int | None
             ) -> SuManageVDBsQueryResult:
         # update list of `api_groups`.`group_id` and `api_groups`.`group_name`
-        manage_vdbs__qc, manage_vdbs__to = await get_qc_to_with_group_id_name(
-            self.session, su_manage_vdbs__query_columns, su_manage_vdbs__table_options
+        manage_vdbs__qc, manage_vdbs__to = await get_qc_to_with_user_id_name_group_id_name(
+            self.session, su_manage_vdbs__query_columns, su_manage_vdbs__table_options,
+            {'group_id': ('add_values', 'select_default')}
         )
         # 
         where_clause = GroupVDBs.gvdbs_id > -1