+++ frontend/src/pages/SuManageLogCRUD/SuManageLogCRUDPage.tsx
@@ -1,0 +1,11 @@
+import { useTopNavBarTitle } from "../../hooks/useTopNavBarTitle";
+import TableSuManageLogCRUD from "./SuTableManageLogCRUD";
+
+export default function SuManageLogCRUDPage() {
+  useTopNavBarTitle("Manage Log CRUD");
+  return (
+    <div className="container-fluid my-4">
+      <TableSuManageLogCRUD />
+    </div>
+  );
+}